teachers/serializers.py
from rest_framework import serializers
from teachers.models import Teacher, Subject


class SubjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Subject
        fields = '__all__'


class TeacherSerializer(serializers.ModelSerializer):
    subjects = SubjectSerializer(many=True)

    class Meta:
        model = Teacher
        fields = '__all__'
